refactor: optimize gesture branching logic and cleanup code
- Replace 9 raw parameters in gesture_decide_branch() with GestureBranchParams struct
- Add GestureBindingSet builder gesture_branch_params() for named field initialization
- Extract repetitive cleanup patterns into inline helpers:
  gesture_clear_deferred_tap(), gesture_clear_pending_long_press(),
  gesture_clear_second_finger_state()
- Add finger_has_gesture() inline to eliminate 3 copies of same check
- Remove redundant TS/TP manual switch in handle_tap_up_impl() —
  setup_second_finger_bindings() already resolves bindings per finger role
- Replace manual for-loops over MAX_FINGERS with find_finger() helper
- Invert empty if/else block in touchpad_finger_up() for clarity
- Clean up pre-existing unused variables and redundant assignments

+++ app/src/main/cpp/winlator/gesture/touchscreen.c
@@ -36,7 +36,7 @@ void handle_touchscreen_down(TouchFinger* f, float x, float y, uint64_t time_ms,
     TouchElement* btn = hit_test_element(x, y);
     if (btn && btn->type == ELEM_BUTTON) {
         if (btn->activation_mode == ACTIVATION_TRACK || btn->activation_mode == ACTIVATION_HOVER) {
-            TrackedButtons* tb = &g_state.tracked[f->ptr_id % MAX_FINGERS];
+            TrackedButtons* tb = &g_state.tracked[(uint32_t)f->ptr_id % MAX_FINGERS];
             bool already = false;
             for (int j = 0; j < tb->count; j++) {
                 if (tb->element_indices[j] == (int)(btn - g_state.elements)) { already = true; break; }
@@ -47,7 +47,7 @@ void handle_touchscreen_down(TouchFinger* f, float x, float y, uint64_t time_ms,
                     tb->element_indices[tb->count++] = (int)(btn - g_state.elements);
             }
             if (btn->activation_mode == ACTIVATION_HOVER)
-                g_state.hovered_element_per_ptr[f->ptr_id % MAX_FINGERS] = (int)(btn - g_state.elements);
+                g_state.hovered_element_per_ptr[(uint32_t)f->ptr_id % MAX_FINGERS] = (int)(btn - g_state.elements);
             if (!btn->passthrough_touch) handled = true;
         } else {
             handle_element_down(btn, f->ptr_id, x, y, time_ms, result);
@@ -274,7 +274,7 @@ void handle_touchscreen_move(TouchFinger* f, float x, float y, uint64_t time_ms,
             ActivationMode mode = g_state.element_count > 0 ?
                 g_state.elements[0].activation_mode : ACTIVATION_LOCK;
             if (mode == ACTIVATION_TRACK || mode == ACTIVATION_HOVER) {
-                int pi = f->ptr_id % MAX_FINGERS;
+                int pi = (uint32_t)f->ptr_id % MAX_FINGERS;
                 TrackedButtons* tb = &g_state.tracked[pi];
                 bool already_tracked = false;
                 for (int j = 0; j < tb->count; j++) {
@@ -307,7 +307,7 @@ void handle_touchscreen_move(TouchFinger* f, float x, float y, uint64_t time_ms,
 
     // TRACK/HOVER: tracked button processing runs every move (matches Java handleMoveByMode)
     {
-        int pi = f->ptr_id % MAX_FINGERS;
+        int pi = (uint32_t)f->ptr_id % MAX_FINGERS;
         TrackedButtons* tb = &g_state.tracked[pi];
         if (tb->count > 0) {
             int first_idx = tb->element_indices[0];
@@ -378,7 +378,7 @@ void handle_touchscreen_move(TouchFinger* f, float x, float y, uint64_t time_ms,
 
     // Java handleMoveByMode: !passthrough check on first tracked → skip gesture
     {
-        TrackedButtons* tb = &g_state.tracked[f->ptr_id % MAX_FINGERS];
+        TrackedButtons* tb = &g_state.tracked[(uint32_t)f->ptr_id % MAX_FINGERS];
         if (tb->count > 0) {
             int first_idx = tb->element_indices[0];
             if (first_idx >= 0 && first_idx < g_state.element_count) {
@@ -470,7 +470,7 @@ void handle_touchscreen_move(TouchFinger* f, float x, float y, uint64_t time_ms,
 
 void handle_touchscreen_up(TouchFinger* f, float x, float y, uint64_t time_ms, TouchActionResult* result) {
     // Java handleUpByMode: process tracked buttons FIRST (matches Java order)
-    int pi = f->ptr_id % MAX_FINGERS;
+    int pi = (uint32_t)f->ptr_id % MAX_FINGERS;
     TrackedButtons* tb = &g_state.tracked[pi];
     bool had_tracked = false;
     if (tb->count > 0) {
